requirements.txt
# Web Framework & Server
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
starlette>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0

# Recommendation Engine & Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.5.0
cachetools>=5.3.0

# AI Assistant
google-genai

# PostgreSQL Driver
psycopg2-binary>=2.9.0
